desktop_app/requirements.txt
paho-mqtt
customtkinter
matplotlib